[Add] DeleteUsersTest - guests_cannot_delete_users_profiles

+++ routes/api.php
@@ -20,7 +20,7 @@ Route::prefix('/auth')->group(function () {
 
 Route::prefix('/users')->group(function () {
     Route::get('/{user}', 'UserController@show')->name('users.show');
-    Route::delete('/{user}', 'UserController@destroy');
+    Route::delete('/{user}', 'UserController@destroy')->middleware('auth:api');
 
     Route::get('/{user}/tracks', 'UserTracksController@index')->name('users.tracks');
     Route::get('/{user}/playlists', 'UserPlaylistsController@index')->name('users.playlists');
